Fix controls example to align with text and to be syntactically correct
The `foaf:primaryTopic` triple was inserted in the middle of another triple so needs extracting to the end of the dataset. However I think the object of the triple was incorrect too. There are 2 related sections of text:
> The control form MUST be attached to the dataset since the form filters the dataset, and not the fragment. The fragment MUST be declared as a subset of the dataset such that their relation becomes apparent.

This is a little ambiguous as the mode of attachment is not specified but was perhaps interpreted as
```
<http://example.org/example#controls> foaf:primaryTopic <http://example.org/example#dataset> .
```
However I think that "control form" in this context actually refers to the `hydra:search` node.

The more relevant text is:
> If the RDF syntax supports multiple graphs, control triples MUST be serialized to a non-default graph. This non-default graph MUST be explicitly related to the Quad Pattern Fragment using the foaf:primaryTopic predicate, so clients can interpret what resource this metadata belongs to. This relating triple must be present in controls graph. This graph SHOULD be the same as the graph containing the metadata.

The correct representation of this would appear to be:
```
<http://example.org/example#controls> foaf:primaryTopic <http://example.org/example?s=http%3A%2F%2Fexample.org%2Ftopic> .
```